fix(backends): widen multiplicative allocations to size_t (tape)
T2.2 from the tier-2 clang-tidy expansion plan. Enables
`bugprone-implicit-widening-of-multiplication-result`, which
flagged 70 sites across 17 files where `int * int` (or longer
chains) get implicitly widened to `size_t` for malloc/arena_alloc
sizing or to `ptrdiff_t` for pointer offsets. Real overflow risk
for shape-product > 2³¹.

Fix pattern: cast the first int multiplicand to `(size_t)` so the
remainder of the chain propagates in size_t. Two common shapes:
  - `malloc(m * n * sizeof(T))` → `malloc((size_t)m * n * sizeof(T))`
  - `data[i * n + j]` → `data[(size_t)i * n + j]`

Touched (all under packages/backends/backend_tape/):
  conv/conv2d_batched.c            (3 sites)
  core/lifecycle/lifecycle_ext.c   (2 sites)
  linear/concat/concat_2d_axis1.c  (8 sites)
  linear/linalg/bmm.c              (6 sites)
  linear/linalg/bmm_3x3.c          (8 sites)
  linear/linalg/linear_2d.c        (9 sites)
  linear/linalg/mm.c               (6 sites)
  linear/linalg/outer.c            (4 sites)
  linear/linalg/tile_2d.c          (4 sites)
  linear/linalg/transpose_2d.c     (4 sites)
  nn/activation/swiglu_2d.c        (4 sites)
  nn/attention/embedding.c         (2 sites)
  nn/mask/expand_mask.c            (4 sites)
  nn/norm/layer_norm_2d.c          (4 sites)
  nn/norm/rms_norm_2d.c            (4 sites)
  nn/softmax/log_softmax_2d.c      (1 site)
  nn/softmax/softmax_2d.c          (1 site)

No semantic change for tensors with shape-product < 2³¹ (every
test the repo currently runs). The change is purely defensive
against larger workloads that the codebase could grow into.

Verification: `make lint-c-tape` exits 0 with the new check
enabled; `make BACKEND=tape backend && make test-unit-c-tape`
passes (235/235 unit tests).

Scope caveat: this only touches backend_tape. Torch and mlx have
parallel implementations of the same ops that almost certainly
contain analogous patterns, but their clang-tidy step is
env-gated off (`C_LINT_FULL_CLANG_TIDY=1`) due to libtorch
header / nix-libc++ vs Apple-SDK incompatibility. CI runs
clang-tidy only on tape, matching this scope. A follow-up commit
can mirror these widening fixes to torch + mlx; for now the
ports remain unchecked.

+++ packages/backends/backend_tape/conv/conv2d_batched.c
@@ -45,7 +45,7 @@ static void conv2d_im2col(const double* input, int B, int inC, int H, int W, int
 		const double* inp_b = input + (size_t)b * inC * H * W;
 		for (int oh = 0; oh < oH; oh++) {
 			for (int ow = 0; ow < oW; ow++) {
-				double* row = X_col + ((size_t)b * oH * oW + oh * oW + ow) * K;
+				double* row = X_col + ((size_t)b * oH * oW + (size_t)oh * oW + ow) * K;
 				for (int ic = 0; ic < inC; ic++) {
 					for (int kh = 0; kh < kH; kh++) {
 						int ih = oh * strideH - padH + kh;
@@ -74,7 +74,7 @@ static void conv2d_im2col_f32(const float* input, int B, int inC, int H, int W, 
 		const float* inp_b = input + (size_t)b * inC * H * W;
 		for (int oh = 0; oh < oH; oh++) {
 			for (int ow = 0; ow < oW; ow++) {
-				float* row = X_col + ((size_t)b * oH * oW + oh * oW + ow) * K;
+				float* row = X_col + ((size_t)b * oH * oW + (size_t)oh * oW + ow) * K;
 				for (int ic = 0; ic < inC; ic++) {
 					for (int kh = 0; kh < kH; kh++) {
 						int ih = oh * strideH - padH + kh;
@@ -101,7 +101,7 @@ static void conv2d_col2im_accumulate(const double* dX_col, int B, int inC, int H
 		double* din_b = dInput + (size_t)b * inC * H * W;
 		for (int oh = 0; oh < oH; oh++) {
 			for (int ow = 0; ow < oW; ow++) {
-				const double* row = dX_col + ((size_t)b * oH * oW + oh * oW + ow) * K;
+				const double* row = dX_col + ((size_t)b * oH * oW + (size_t)oh * oW + ow) * K;
 				for (int ic = 0; ic < inC; ic++) {
 					for (int kh = 0; kh < kH; kh++) {
 						int ih = oh * strideH - padH + kh;
